review comments taken into account
